04.10.24edutgapp - Copy/pyproject.toml
[tool.poetry]
name = "edutgapp"
version = "0.1.0"
description = ""
authors = ["Riakol <inspi47@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram-dialog = "^2.2.0"
python-decouple = "^3.8"
asyncpg = "^0.29.0"
asyncio = "^3.4.3"
pandas = "^2.2.3"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
